pyproject.toml
[tool.poetry]
name = "llmexplorer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.2"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.21"
psycopg2 = "^2.9.8"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.3"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-mail = "^1.4.1"
libgravatar = "^1.0.4"
slugify = "^0.0.1"
alembic = "^1.12.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
python-slugify = "^8.0.1"
streamlit = "^1.27.2"
pypdf = "^3.16.2"
langchain = "^0.0.309"
pypdf2 = "^3.0.1"
translate = "^3.6.1"
openai = "^0.28.1"
tiktoken = "^0.5.1"
faiss-cpu = "^1.7.4"
pip = "^23.2.1"
install = "^1.3.5"
jinja2 = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"